+++ index.ts
@@ -2,6 +2,7 @@ export const version: string = require("../package.json").version as string;
 
 export * from "./src/@Managers/MoonlinkManager";
 export * from "./src/@Managers/PlayerManager";
+export * from "./src/@Managers/NodeManager";
 export * from "./src/@Entities/MoonlinkNode";
 export * from "./src/@Entities/MoonlinkPlayer";
 export * from "./src/@Entities/MoonlinkQueue";
