+++ src/@Typings/index.ts
@@ -8,8 +8,8 @@ import {
     MoonlinkNode,
     MoonlinkTrack,
     PlayerManager,
-    Plugin,
-    Nodes
+    NodeManager,
+    Plugin
 } from "../../index";
 
 export type Constructor<T> = new (...args: any[]) => T;
@@ -158,7 +158,7 @@ export interface Extendable {
     MoonlinkNode: typeof MoonlinkNode;
     MoonlinkTrack: typeof MoonlinkTrack;
     PlayerManager: typeof PlayerManager;
-    Nodes: typeof Nodes;
+    NodeManager: typeof NodeManager;
 }
 
 export interface PlaylistInfo {
